requirements.txt
asgiref==3.5.1
Django==4.0.4
djangorestframework==3.13.1
httplib2==0.20.4
Pillow==9.1.0
pyparsing==3.0.8
pytz==2022.1
sqlparse==0.4.2
requests==2.27.1
pycryptodome>=3.14.1
cryptography>=36.0.2
